Rename util.js to apiUtil.js
