+++ backend/busController.js
@@ -1,6 +1,6 @@
 import axios from "axios";
 import { xml2js } from "xml-js";
-import { formatTextProperty } from "./util";
+import { formatTextProperty } from "./apiUtil";
 
 const ROOT = "http://www.bt4uclassic.org/webservices/bt4u_webservice.asmx";
